Update mpool_replace.py
Alterations to actually be operating on "epochs" well close to them. We get the current epoch on startup and then equalize to 30 second intervals but at any one time if you were to compare script epoch to chain epoch it should be within a hairline tolerance.

Lots of factors can cause 1 to 2 second delays for lotus to be on the exact network epoch.